Refactor: Extract `ChannelMediaAttachmentsList`
This commit extracts the `ChannelMediaAttachmentsList` composable from `ChannelMediaAttachmentsScreen.kt` into its own file. It also renames the existing screen scaffold composable from `ChannelMediaAttachmentsScaffold` to `ChannelMediaAttachmentsContent` for clarity.

The extracted list composable is responsible for displaying the grid of media attachments, handling loading states, empty states, error states, and pagination.

Additionally, the size of the play icon in `ChannelMediaAttachmentsItem` is reduced from 0.3f to 0.25f.
